// src/app/core/interfaces/productoEmpaque.interface.ts
export interface ProductoEmpaque {
  idProductoEmpaque:       number;
  idProducto:              number;
  descripcionProducto:     string;
  idPresentacion:          number;
  descripcionPresentacion: string;
  cantidad:                number;
  precioVenta:             string;
  precioCompra:            string;
  esActivo:                number;
}